+++ rvp64_scan/builtin-rvp64-sra16.c
@@ -9,7 +9,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-uint32_t ars (uint32_t ra, uint32_t rb)
+uint64_t ars (uint64_t ra, uint32_t rb)
 {
   return __rv__sra16 (ra, rb);
 }
